1. Add printing an order on a pdf file 2. Add filtration of companies by name and income

+++ project source/src/main/java/org/example/dao/AbstractDao.java
@@ -26,7 +26,7 @@ public abstract class AbstractDao<T> implements CrudDao<T> {
         }
     }
 
-    public T getById(long id) {
+    public T getById(int id) {
         T entity;
         try (Session session = HibernateConfig.getSessionFactory().openSession()) {
             Transaction transaction = session.beginTransaction();
